optimize2 auth action&reducer code, no new function/feature
