+++ src/EventHive/Auth/client.js
@@ -1,25 +1,35 @@
 import axios from '../../config/axiosConfig';
-
 import { showAlert } from '../../utils/alertHelper';
 import { userLoaded, userAuthenticated, userNotAuthenticated } from './AuthReducer';
 
+const generateConfig = () => ({
+	headers: {
+	  'Content-Type': 'application/json',
+	},
+  });
+
+  const handleErrorResponse = (err) => {
+	const errors = err.response.data.errors;
+  
+	if (errors) {
+	  showAlert('error', 'Oops...', errors.map((error) => error.msg).join('\n'));
+	}
+  };
+
 export const loadUser = () => async (dispatch) => {
-	console.log("begin loadUser()============");
 
 	if (localStorage.apiKey) {
 		axios.defaults.headers.common['x-api-key'] = localStorage.apiKey;
 	} else {
 		delete axios.defaults.headers.common['x-api-key'];
 	}
-	
+
 	try {
 		const res = await axios.get('/users/load');
 
-		// dispatch({ type: 'USER_LOADED', payload: res.data });
 		dispatch(userLoaded(res.data));
 
 	} catch (err) {
-		// dispatch({ type: 'NotAuthenticated' });
 		dispatch(userNotAuthenticated());
 	}
 };
@@ -27,22 +37,11 @@ export const loadUser = () => async (dispatch) => {
 export const register =
 	({ name, email, password }) =>
 	async (dispatch) => {
-		const config = {
-			headers: {
-				'Content-Type': 'application/json',
-			},
-		};
+
 		const body = JSON.stringify({ name, email, password });
 
 		try {
-			const res = await axios.post('/users/register', body, config);
-
-			// dispatch({
-			// 	type: 'Authenticated',
-			// 	payload: res.data,
-			// });
-			// dispatch(loadUser());
-
+			const res = await axios.post('/users/register', body, generateConfig());
 
 			dispatch(userAuthenticated(res.data));
 			dispatch(loadUser());
@@ -50,66 +49,57 @@ export const register =
 
 			showAlert('success', 'Success!', 'SignUp Successfully');
 		} catch (err) {
-			const errors = err.response.data.errors;
-
-			if (errors) {
-				showAlert(
-					'error',
-					'Oops...',
-					errors.map((error) => error.msg).join('<br/>')
-				);
-			}
-
-			// dispatch({
-			// 	type: 'NotAuthenticated',
-			// });
-
+			handleErrorResponse(err);
 			dispatch(userNotAuthenticated());
 		}
 	};
 
-// Login user
 export const login = (email, password, isAdministrator) => async (dispatch) => {
-	const config = {
-		headers: {
-			'Content-Type': 'application/json',
-		},
-	};
 
 	const body = JSON.stringify({ email, password, isAdministrator });
 
 	try {
-		const res = await axios.post('/users/login', body, config);
-		// dispatch({
-		// 	type: 'Authenticated',
-		// 	payload: res.data,
-		// });
-
-		// dispatch(loadUser());
+		const res = await axios.post('/users/login', body, generateConfig());
 
 		dispatch(userAuthenticated(res.data));
 		dispatch(loadUser());
 
 		showAlert('success', 'Success!', 'Login Successfully');
 	} catch (err) {
-		const errors = err.response.data.errors;
-
-		if (errors) {
-			showAlert(
-				'error',
-				'Oops...',
-				errors.map((error) => error.msg).join('<br/>')
-			);
-		}
-		// dispatch({
-		// 	type: 'NotAuthenticated',
-		// });
+		handleErrorResponse(err);
 		dispatch(userNotAuthenticated());
 	}
 };
 
 export const logout = () => (dispatch) => {
-	// dispatch({ type: 'NotAuthenticated' });
 	dispatch(userNotAuthenticated());
 	showAlert('success', 'Success!', 'Logout Successfully');
 };
+
+export const resetPassword = async (email, validationCode, password, navigate) => {
+
+	const body = JSON.stringify({email, validationCode, password});
+
+	try {
+		const res = await axios.post('/users/reset-password', body, generateConfig());
+
+		showAlert('success', 'Success!', 'Reset Successfully, Jump into Login Page');
+		navigate('/EventHive/login');
+
+	} catch (err) {
+		handleErrorResponse(err);
+	}
+}
+
+export const sendValidationCode = async (email) => {
+
+	const body = JSON.stringify({email});
+
+	try {
+		const res = await axios.post('/users/forgot-password', body, generateConfig());
+
+		showAlert('success', 'Success!', `Validation code sent to ${email}`);
+	} catch (err) {
+		handleErrorResponse(err);
+	}
+}
